+++ sdk/data/azcosmos/cosmos_http_constants.go
@@ -91,6 +91,7 @@ const (
 	headerDedicatedGatewayBypassCache              string = "x-ms-dedicatedgateway-bypass-cache"
 	cosmosHeaderPriorityLevel                      string = "x-ms-cosmos-priority-level"
 	cosmosHeaderThroughputBucket                   string = "x-ms-cosmos-throughput-bucket"
+	cosmosHeaderRetryAfterMs                       string = "x-ms-retry-after-ms"
 )
 
 const (